+++ nym-registration-client/src/clients/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2026 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod lp;
+pub(crate) mod mixnet;
+
+pub use lp::LpBasedRegistrationClient;
+pub use mixnet::MixnetBasedRegistrationClient;
